Move InChI related methods from generator/parser to inchi module
