+++ rmgpy/molecule/parser.py
@@ -49,14 +49,10 @@ except :
 
 from rdkit import Chem
 
-from rmgpy.molecule import element as elements
-from .molecule import Atom, Bond, Molecule
-from .adjlist import ConsistencyChecker
 from rmgpy.molecule.converter import fromRDKitMol, fromOBMol
 
 import rmgpy.molecule.inchi as inchiutil
 import rmgpy.molecule.util as util
-import rmgpy.molecule.pathfinder as pathfinder
 
 # constants
 
@@ -224,83 +220,6 @@ def __lookup(mol, identifier, type_identifier):
         except KeyError:
             return None
 
-def check(mol, aug_inchi):
-    """
-    Check if the molecular structure is correct.
-
-    Checks whether the multiplicity contained in the augmented inchi, 
-    corresponds to the number of unpaired electrons + 1 found in the molecule.
-
-    Checks whether the valence of each atom is compatible with the bond order,
-    number of unpaired electrons, lone pairs and charge.
-
-    """
-    cython.declare(inchi=str,
-                   at=Atom
-                   )
-
-    ConsistencyChecker.check_multiplicity(mol.getRadicalCount(), mol.multiplicity)
-    inchi, u_indices, p_indices = inchiutil.decompose(str(aug_inchi))
-    assert(mol.getRadicalCount() == len(u_indices))
-    
-    for at in mol.atoms:
-        ConsistencyChecker.check_partial_charge(at)
-
-def fix_oxygen_unsaturated_bond(mol, u_indices):
-    """
-    Searches for a radical or a charged oxygen atom connected to 
-    a closed-shell carbon via an unsatured bond.
-
-    Decrements the unsatured bond,
-    transfers the unpaired electron from O to C or
-    converts the charge from O to an unpaired electron on C, 
-    increases the lone pair count of O to 2.
-
-    Only do this once per molecule.
-    """
-
-    for at in mol.atoms:
-        if at.isOxygen() and at.radicalElectrons == 1 and at.lonePairs == 1:
-            bonds = mol.getBonds(at)
-            oxygen = at
-            for atom2, bond in bonds.iteritems():
-                if bond.isTriple():
-                    bond.decrementOrder()
-                    oxygen.radicalElectrons -= 1
-                    atom2.radicalElectrons += 1
-                    oxygen.lonePairs += 1
-                    return
-        elif at.isOxygen() and at.charge == 1 and at.lonePairs == 1:
-            bonds = mol.getBonds(at)
-            oxygen = at
-
-            start = oxygen
-            # search for 3-atom-2-bond [X=X-X] paths
-            paths = pathfinder.find_allyl_end_with_charge(start)
-            for path in paths:    
-                end = path[-1]
-                start.charge += 1 if start.charge < 0 else -1
-                end.charge += 1 if end.charge < 0 else -1
-                start.lonePairs += 1
-                # filter bonds from path and convert bond orders:
-                bonds = path[1::2]#odd elements
-                for bond in bonds[::2]:# even bonds
-                    assert isinstance(bond, Bond)
-                    bond.decrementOrder()
-                for bond in bonds[1::2]:# odd bonds
-                    assert isinstance(bond, Bond)
-                    bond.incrementOrder()  
-                return
-            else:
-                for atom2, bond in bonds.iteritems():
-                    if not bond.isSingle() and atom2.charge == 0:
-                        oxygen.charge -= 1
-                        if (mol.atoms.index(atom2) + 1) in u_indices:
-                            bond.decrementOrder()
-                            atom2.radicalElectrons += 1
-                            u_indices.remove(mol.atoms.index(atom2) + 1)
-                        oxygen.lonePairs += 1
-                        return
 
 def fromInChI(mol, inchistr, backend='try-all'):
     """
@@ -340,7 +259,7 @@ def fromAugmentedInChI(mol, aug_inchi):
 
     mol.multiplicity = len(aug_inchi.u_indices) + 1 if aug_inchi.u_indices else 1
 
-    fix(mol, aug_inchi)
+    inchiutil.fix(mol, aug_inchi)
 
     mol.updateAtomTypes()
 
@@ -365,440 +284,3 @@ def fromSMARTS(mol, smartsstr, backend = 'rdkit'):
     return __parse(mol, smartsstr, 'sma', backend)
 
 
-def fixCharge(mol, u_indices):
-    """
-    Tries to fix a number of structural features in the molecule related to charge, 
-    based on the information from the parameter list of atom indices with unpaired electrons.
-    """
-
-    if not u_indices:
-        return
-
-    is_charged = sum([abs(at.charge) for at in mol.atoms]) != 0
-    is_correct = mol.getRadicalCount() == (mol.multiplicity - 1)
-    if mol.multiplicity < 3 or is_correct or not is_charged:
-        return
-
-    # converting charges to unpaired electrons for atoms in the u-layer
-    convert_charge_to_unpaired_electron(mol, u_indices)
-
-    # convert neighboring atoms (or delocalized paths) to unpaired electrons
-    convert_delocalized_charge_to_unpaired_electron(mol, u_indices)
-
-    fix_adjacent_charges(mol)
-
-def check_bond_order_oxygen(mol):
-    """Check if total bond order of oxygen atoms is smaller than 4."""
-    from rmgpy.molecule.util import ORDERS
-
-    for at in mol.atoms:
-        if at.number == 8:
-            order = sum([ORDERS[b.order] for _, b in at.bonds.iteritems()])
-            not_correct = order >= 4
-            if not_correct:
-                return False
-
-    return True
-
-def find_mobile_h_system(mol, all_mobile_h_atoms_couples, test_indices):
-    """
-    
-    """
-    dummy = test_indices[:]
-
-    for mobile_h_atom_couple in all_mobile_h_atoms_couples:
-        for test_index in test_indices:
-            if test_index in mobile_h_atom_couple:
-                original_atom = test_index
-                dummy.remove(test_index)
-                mobile_h_atom_couple.remove(test_index)
-                new_partner = mobile_h_atom_couple[0]
-                central = dummy[0]
-                return mol.atoms[central - 1], mol.atoms[original_atom - 1], mol.atoms[new_partner - 1]
-
-    raise Exception('We should always have found the mobile-H system. All mobile H couples: {}, test indices: {}'
-        .format(all_mobile_h_atoms_couples, test_indices))
-    
-def fix_adjacent_charges(mol):
-    """
-    Searches for pairs of charged atoms.
-    Neutralizes one unit of charge on each atom,
-    and increments the bond order of the bond in between
-    the atoms.
-    """
-    for at in mol.atoms:
-        if at.charge != 0:
-            for neigh, bond in at.bonds.iteritems():
-                if neigh.charge != 0:
-                    bond.incrementOrder()
-                    at.charge += 1 if at.charge < 0 else -1
-                    neigh.charge += 1 if neigh.charge < 0 else -1
-
-def convert_charge_to_unpaired_electron(mol, u_indices):
-    """
-    Iterates over the atoms foundin the parameter list and
-    converts a unit of charge on atoms into an unpaired electron.
-
-    Removes treated atoms from the parameter list.
-    """
-    for at in mol.atoms:
-        at_index = mol.atoms.index(at) + 1
-        if at.charge != 0 and at_index in u_indices:
-            at.charge += 1 if at.charge < 0 else -1
-            at.radicalElectrons += 1
-            u_indices.remove(at_index)                    
-
-def convert_delocalized_charge_to_unpaired_electron(mol, u_indices):
-    """
-    Iterates over the atom indices of the parameter list and searches 
-    a charged atom that is connected to that atom via some kind of
-    delocalization path.
-
-    """
-    u_indices_copy = u_indices[:]
-    for index in u_indices_copy:
-        start = mol.atoms[index -1]
-
-        found = convert_4_atom_3_bond_path(start)
-        if found: 
-            u_indices.remove(index)
-            continue
-
-        found = convert_3_atom_2_bond_path(start, mol)
-        if found:
-            u_indices.remove(index)
-            continue
-
-def convert_4_atom_3_bond_path(start):
-    """
-    Searches for 4-atom-3-bond [X=X-X=X+] paths starting from the parameter atom.
-    If a path is found, the starting atom receives an unpaired electron while
-    the bonds in the delocalization path are "inverted". A unit of charge on the 
-    end atom is neutralized and a lone pair is added.
-    """
-    path = pathfinder.find_butadiene_end_with_charge(start)
-
-    if path is not None:    
-        start.radicalElectrons += 1
-        end = path[-1]
-        end.charge += 1 if end.charge < 0 else -1
-        end.lonePairs += 1
-
-        # filter bonds from path and convert bond orders:
-        bonds = path[1::2]#odd
-        for bond in bonds[::2]:# even
-            assert isinstance(bond, Bond)
-            bond.decrementOrder()
-        for bond in bonds[1::2]:# odd bonds
-            assert isinstance(bond, Bond)
-            bond.incrementOrder()  
-
-        return True
-
-    return False
-
-def convert_3_atom_2_bond_path(start, mol):
-    """
-    Searches for 3-atom-2-bond [X=X-X+] paths paths starting from the parameter atom.
-    If a correct path is found, the starting atom receives an unpaired electron while
-    the bonds in the delocalization path are "inverted". A unit of charge on the 
-    end atom is neutralized and a lone pair is added.
-
-    If it turns out the path was invalid, the actions are reverted, and another path
-    is tried instead.
-
-    To facilitate reverting the changes, we use a reaction recipe and populate it
-    with a number of actions that reflect the changes in bond orders and unpaired
-    electrons that the molecule should undergo.
-    """
-    from rmgpy.data.kinetics.family import ReactionRecipe
-
-    def is_valid(mol):
-        """Check if total bond order of oxygen atoms is smaller than 4."""
-
-        for at in mol.atoms:
-            if at.number == 8:
-                order = at.getBondOrdersForAtom()
-                not_correct = order >= 4
-                if not_correct:
-                    return False
-
-        return True
-
-    index = mol.atoms.index(start) + 1
-
-    paths = pathfinder.find_allyl_end_with_charge(start)
-
-    for path in paths:
-        # label atoms so that we can use the labels in the actions of the recipe
-        for i, at in enumerate(path[::2]):
-            at.label = str(i)
-        # we have found the atom we are looking for
-        recipe = ReactionRecipe()
-        recipe.addAction(['GAIN_RADICAL', start.label, 1])
-
-        end = path[-1]
-        end_original_charge = end.charge
-      
-        # filter bonds from path and convert bond orders:
-        bonds = path[1::2]#odd elements
-        for bond in bonds[::2]:# even
-            recipe.addAction(['CHANGE_BOND', bond.atom1.label, -1, bond.atom2.label])
-        for bond in bonds[1::2]:# odd
-            recipe.addAction(['CHANGE_BOND', bond.atom1.label, 1, bond.atom2.label])
-
-        end.charge += 1 if end.charge < 0 else -1
-        recipe.applyForward(mol)
-
-        if is_valid(mol):
-            # unlabel atoms so that they never cause trouble downstream
-            for i, at in enumerate(path[::2]):
-                at.label = ''
-            return True
-        else:
-            recipe.applyReverse(mol)
-            end.charge = end_original_charge
-
-            # unlabel atoms so that they never cause trouble downstream
-            for i, at in enumerate(path[::2]):
-                assert isinstance(at, Atom)
-                at.label = ''
-
-    return False
-
-def fix(mol, aug_inchi):
-    """
-    Fixes a number of structural features of the erroneous Molecule
-    parsed by the backends, based on multiplicity and unpaired electron information
-    stored in the augmented inchi.
-    """   
-
-    u_indices = aug_inchi.u_indices[:] if aug_inchi.u_indices else []
-    p_indices = aug_inchi.p_indices[:] if aug_inchi.p_indices else []
-
-    # ignore atoms that bear already unpaired electrons:
-    for i in set(u_indices[:]):
-        atom = mol.atoms[i - 1]
-        [u_indices.remove(i) for _ in range(atom.radicalElectrons)]        
-
-    # ignore atoms that bear already lone pairs:
-    for i in set(p_indices[:]):
-        atom = mol.atoms[i - 1]
-        [p_indices.remove(i) for _ in range(atom.lonePairs)]   
-
-
-    fix_triplet_to_singlet(mol, p_indices)
-    
-    fixCharge(mol, u_indices)
-                                
-    reset_lone_pairs(mol, p_indices)
-
-    fix_oxygen_unsaturated_bond(mol, u_indices)
-
-    fix_unsaturated_bond(mol, u_indices, aug_inchi)
-
-    check(mol, aug_inchi)    
-
-
-def fix_triplet_to_singlet(mol, p_indices):
-    """
-    Iterates over the atoms and checks whether atoms bearing two unpaired electrons are
-    also present in the p_indices list.
-
-    If so, convert to the two unpaired electrons into a lone pair, and remove that atom
-    index from the p_indices list.
-    """
-
-    for at in mol.atoms:
-        index = mol.atoms.index(at) + 1
-        if mol.getRadicalCount() == 2 and index in p_indices:
-            at.lonePairs += 1
-            at.radicalElectrons -= 2
-            p_indices.remove(index)
-
-
-def fix_butadiene_path(start, end):
-    """
-    Searches for a 1,3-butadiene path between the start and end atom.
-    Adds an unpaired electron to start and end atom, and "inverts" the bonds
-    in between them. 
-    """
-    path = pathfinder.find_butadiene(start, end)
-    if path is not None:
-        start.radicalElectrons += 1
-        end.radicalElectrons += 1
-        # filter bonds from path and convert bond orders:
-        bonds = path[1::2]#odd elements
-        for bond in bonds[::2]:# even bonds
-            assert isinstance(bond, Bond)
-            bond.decrementOrder()
-        for bond in bonds[1::2]:# odd bonds
-            assert isinstance(bond, Bond)
-            bond.incrementOrder()    
-
-        return True
-
-    return False
-
-def fix_mobile_h(mol, inchi, u1, u2):
-    """
-
-    Identifies a system of atoms bearing unpaired electrons and mobile hydrogens
-    at the same time.
-
-    The system will consist of a central atom that does not bear any mobile hydrogens,
-    but that is bound to an atom that does bear a mobile hydrogen, called the "original atom".
-
-    The algorithm identifies the "new partner" atom that is part of the mobile hydrogen 
-    system.
-
-    Next, the mobile hydrogen is transferred from the original atom, to the new partner,
-    and a bond is removed and added respectively.
-
-    Finally, the central atom and the original atom will each receive an unpaired electron,
-    and the bond between them will decrease in order.
-    """
-
-    mobile_hydrogens = inchiutil.parse_H_layer(inchi)
-
-    if mobile_hydrogens:
-        # WIP: only consider the first system of mobile hydrogens:
-        mobile_hydrogens = mobile_hydrogens[0]
-
-        #find central atom:
-        central, original, new_partner = util.swap(mobile_hydrogens, [u1, u2])
-
-        central, original, new_partner = \
-        mol.atoms[central - 1], mol.atoms[original - 1], mol.atoms[new_partner - 1]
-
-        # search hydrogen atom and bond
-        hydrogen = None
-        for at, bond in original.bonds.iteritems():
-            if at.number == 1:
-                hydrogen = at
-                mol.removeBond(bond)
-                break
-
-        new_h_bond = Bond(new_partner, hydrogen, order='S')
-        mol.addBond(new_h_bond)
-        
-        mol.getBond(central, new_partner).decrementOrder()
-
-        central.radicalElectrons += 1
-        original.radicalElectrons += 1
-        return True
-
-    return False
-
-def convert_unsaturated_bond_to_triplet(bond):
-    """
-    Decrements the bond if it is unsatured, and adds an unpaired
-    electron to each of the atoms connected by the bond.
-    """
-    if not bond.isSingle():
-        for at in (bond.atom1, bond.atom2):
-            at.radicalElectrons += 1
-        bond.decrementOrder()    
-        return True
-    return False
-
-def reset_lone_pairs(mol, p_indices):
-    """
-    Iterates over the atoms of the molecule and
-    resets the atom's lone pair count to the value stored in the p_indices list,
-    or to the default value.
-
-    """
-    for at in mol.atoms:
-        index = mol.atoms.index(at) + 1 #1-based index
-        count = p_indices.count(index)
-        if count != 0:
-            at.lonePairs = count
-        else:    
-            order = at.getBondOrdersForAtom()
-            at.lonePairs = (elements.PeriodicSystem.valence_electrons[at.symbol] - order - at.radicalElectrons - at.charge) / 2
-
-def fix_unsaturated_bond_to_biradical(mol, inchi, u_indices):
-    """
-    Convert an unsaturated bond (double, triple) into a bond
-    with a lower bond order (single, double), and give an unpaired electron
-    to each of the neighboring atoms, with indices referring to the 1-based
-    index in the InChI string.
-    """
-    cython.declare(u1=cython.int, u2=cython.int)
-    cython.declare(atom1=Atom, atom2=Atom)
-    cython.declare(b=Bond)
-
-    combos = itertools.combinations(u_indices, 2)
-
-    isFixed = False
-    for u1, u2 in combos:
-        atom1 = mol.atoms[u1 - 1] # convert to 0-based index for atoms in molecule
-        atom2 = mol.atoms[u2 - 1] # convert to 0-based index for atoms in molecule
-        if mol.hasBond(atom1, atom2):
-            b = mol.getBond(atom1, atom2)
-            isFixed = convert_unsaturated_bond_to_triplet(b)
-            if isFixed:
-                break
-                
-            else:
-                isFixed = fix_mobile_h(mol, inchi, u1, u2)
-                if isFixed:
-                    break
-        else:
-            isFixed = fix_butadiene_path(atom1, atom2)
-            if isFixed:
-                break
-
-    if isFixed:
-        u_indices.remove(u1)
-        u_indices.remove(u2)
-        return mol                
-    else:
-        raise Exception(
-            'Could not convert an unsaturated bond into a biradical for the \
-            indices {} provided in the molecule: {}.'
-            .format(u_indices, mol.toAdjacencyList())
-            )    
-
-def isUnsaturated(mol):
-    """
-    Does the molecule have a bond that's not single?
-    Eg. a bond that is double or triple or benzene
-    """
-    cython.declare(atom1=Atom,
-                   atom2=Atom,
-                   bonds=dict,
-                   bond=Bond)
-    for atom1 in mol.atoms:
-        bonds = mol.getBonds(atom1)
-        for atom2, bond in bonds.iteritems():
-            if not bond.isSingle():
-                return True
-
-    return False
-
-
-def fix_unsaturated_bond(mol, indices, aug_inchi):
-    """
-    Adds unpaired electrons to the molecule by converting unsaturated bonds into triplets.
-
-    It does so by converting an unsaturated bond into a triplet, and verifying whether
-    the total number of unpaired electrons matches the multiplicity.
-
-    Finishes when all unsaturated bonds have been tried, or when there are no pairs
-    of atoms that should be unpaired electrons left. 
-    """
-
-    correct = mol.getRadicalCount() == (mol.multiplicity - 1)
-    
-    if not correct and not indices:
-        raise Exception( 'Cannot correct {} based on {} by converting unsaturated bonds into unpaired electrons...'\
-            .format(mol.toAdjacencyList(), aug_inchi))
-
-    unsaturated = isUnsaturated(mol)
-
-    while not correct and unsaturated and len(indices) > 1:
-        mol = fix_unsaturated_bond_to_biradical(mol, aug_inchi.inchi, indices)
-        correct = mol.getRadicalCount() == (mol.multiplicity - 1)
-        unsaturated = isUnsaturated(mol)
